@@ -22,11 +22,11 @@
     }
   ],
   "require": {
-    "php": ">=5.3.0",
-    "voku/portable-utf8": "~3.1"
+    "php": ">=7.0.0",
+    "voku/portable-utf8": "~4.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "~4.0|~5.0"
+    "phpunit/phpunit": "~6.0"
   },
   "autoload": {
     "psr-4": {
@@ -35,7 +35,7 @@
   },
   "extra": {
     "branch-alias": {
-      "dev-master": "2.0.x-dev"
+      "dev-master": "3.0.x-dev"
     }
   }
 }
