@@ -3,4 +3,4 @@
 error_reporting(E_ALL);
 ini_set('display_errors', 1);
 
-require_once dirname(__DIR__) . '/vendor/composer/autoload_real.php';
+require_once dirname(__DIR__) . '/vendor/autoload.php';
